index iterator, data adjusted types

+++ internal/engine/data/data.go
@@ -24,7 +24,7 @@ func New(path string) (*Data, error) {
 }
 
 // Read reads data from storage
-func (d *Data) Read(offset int64, length int) ([]byte, error) {
+func (d *Data) Read(offset int64, length int32) ([]byte, error) {
 	reader, err := d.storage.Reader(offset)
 	if err != nil {
 		return nil, err
@@ -36,14 +36,14 @@ func (d *Data) Read(offset int64, length int) ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
-	if n != length {
+	if int32(n) != length {
 		return nil, fmt.Errorf("expected %d bytes, got %d", length, n)
 	}
 	return buf, nil
 }
 
 // Append appends data to storage and returns offset and length
-func (d *Data) Append(data []byte) (int64, int, error) {
+func (d *Data) Append(data []byte) (int64, int32, error) {
 	offset, err := d.storage.Size()
 	if err != nil {
 		return 0, 0, err
@@ -59,5 +59,5 @@ func (d *Data) Append(data []byte) (int64, int, error) {
 	if err != nil {
 		return 0, 0, err
 	}
-	return offset, n, nil
+	return offset, int32(n), nil
 }
